scripts: q, Add commit message template support
Add a dictionary of commit message templates to the script. This allows
users to specify a template via a new command-line argument `-t` or
`--template`. If the specified template is not found, the script will
print an error message and list available templates. The selected
template will be used as the instruction for the request if provided.

Example usage:
```
./q -t commit
```